+++ cordex/core/domain.py
@@ -70,6 +70,15 @@ def cordex_domain(short_name, dummy=False, tables=None, attrs=None):
     Dataset : xarray.core.Dataset
         Dataset containing the coordinates.
 
+    Example
+    -------
+
+    To create a cordex rotated pole domain dataset, you can use ,e.g.,::
+
+        import cordex as cx
+
+        eur11 = cx.cordex_domain('EUR-11')
+
     """
     if attrs is None:
         attrs = {}
@@ -83,8 +92,18 @@ def cordex_domain(short_name, dummy=False, tables=None, attrs=None):
 
 
 def create_dataset(
-    nlon, nlat, dlon, dlat, ll_lon, ll_lat, pollon, pollat, 
-    name=None, dummy=False, attrs=None, **kwargs
+    nlon,
+    nlat,
+    dlon,
+    dlat,
+    ll_lon,
+    ll_lat,
+    pollon,
+    pollat,
+    name=None,
+    dummy=False,
+    attrs=None,
+    **kwargs
 ):
     """Create domain dataset from grid information.
 
@@ -114,12 +133,12 @@ def create_dataset(
         Global attributes that should be added to the dataset. If `attrs='CORDEX'`
         a set of standard CF global attributes.
     """
-    if attrs == 'CORDEX':
-        attrs=cf.DEFAULT_CORDEX_ATTRS
+    if attrs == "CORDEX":
+        attrs = cf.DEFAULT_CORDEX_ATTRS
     elif attrs is None:
         attrs = {}
     if name:
-        attrs['CORDEX_domain'] = name
+        attrs["CORDEX_domain"] = name
     rlon, rlat = _init_grid(nlon, nlat, dlon, dlat, ll_lon, ll_lat)
     lon, lat = rotated_coord_transform(*_stack(rlon, rlat), pollon, pollat)
     pole = _grid_mapping(pollon, pollat)
@@ -317,3 +336,41 @@ def rotated_coord_transform(lon, lat, np_lon, np_lat, direction="rot2geo"):
     lat_new = np.rad2deg(lat_new)
 
     return lon_new, lat_new
+
+
+def map_crs(lon, lat, projection, transform=None):
+    """coordinate transformation of longitude and latitude
+
+    Transforms the coordinates lat, lon from the transform crs
+    into the projection crs using cartopy.crs.
+
+    Parameters
+    ----------
+    lon : float array like
+        Longitude coordinate.
+    lat : float array like
+        Latitude coordinate.
+    projection : cartopy.crs
+        Target coordinate reference system into which lat and lon
+        should be projected.
+    transform : cartopy.crs
+        Source coordinate reference system in which lat and lon
+        are defined.
+
+    Returns
+    -------
+    lon : array like
+        Projected longitude coordinate.
+    lat : array like
+        Projected latitude coordinate.
+
+    """
+
+    from cartopy import crs as ccrs
+
+    if transform is None:
+        transform = ccrs.PlateCarree()
+    latlon = ccrs.PlateCarree()
+    lon_stack, lat_stack = xr.broadcast(lon, lat)
+    result = latlon.transform_points(projection, lon_stack.values, lat_stack.values)
+    return result[:, :, 0], result[:, :, 1]