+++ cordex/tables/__init__.py
@@ -15,7 +15,7 @@ class read_cls:
     def table(self):
         return pd.concat(self.tables.values())
 
-    #def __getattr__(self, table):
+    # def __getattr__(self, table):
     #    return self.tables[table]